Separated string manipulation helper functions to one file
